Prepare github release CLI
