+++ lib/github-interface.js
@@ -170,6 +170,8 @@ async function tagRepo ({ repo, version }) {
 
 // Adds a release (with the prerelease flag) to the supplied version, with the supplied message, to the given repo
 function prereleaseRepo ({ repo, version, message }) {
+  if (!semverRegex().test(version)) return new Error('version must be a valid semver')
+  
   return request({
     method: 'POST',
     url: `${githubApiBase}/repos/${repo}/releases`,
